Split out `single_match`
